fix: update some files

+++ internal/entity/add_cart_item.go
@@ -6,8 +6,8 @@ type AddItemRequest struct {
 }
 
 type PriceResponse struct {
-	CartID          int     `json:"cart_id"`
+	CartID          int64   `json:"cart_id"`
 	TotalPrice      float64 `json:"total_price"`
-	DiscountPercent int     `json:"discount_percent"`
+	DiscountPercent int64   `json:"discount_percent"`
 	FinalPrice      float64 `json:"final_price"`
 }
