+++ internal/service/repository.go
@@ -9,7 +9,7 @@ import (
 //go:generate mockery --name=Repository --output=mocks --outpkg=mocks --filename=mock_repository.go --with-expecter
 type Repository interface {
 	CreateCart(ctx context.Context) (*entity.Cart, error)
-	AddCartItem(ctx context.Context, cartID int, product string, price float64) (*entity.CartItem, error)
-	RemoveCartItem(ctx context.Context, cartID int, cartItemID int) error
-	GetCart(ctx context.Context, cartID int) (*entity.Cart, error)
+	AddCartItem(ctx context.Context, cartID int64, product string, price float64) (*entity.CartItem, error)
+	RemoveCartItem(ctx context.Context, cartID int64, cartItemID int64) error
+	GetCart(ctx context.Context, cartID int64) (*entity.Cart, error)
 }
