+++ internal/service/servicer.go
@@ -8,8 +8,8 @@ import (
 
 type Servicer interface {
 	CreateCart(ctx context.Context) (*entity.Cart, error)
-	AddCartItemToCart(ctx context.Context, cartID int, product string, price float64) (*entity.CartItem, error)
-	RemoveItem(ctx context.Context, cartID int, cartItemID int) error
-	GetCart(ctx context.Context, cartID int) (*entity.Cart, error)
-	CalculatePrice(ctx context.Context, cartID int) (*entity.PriceResponse, error)
+	AddCartItemToCart(ctx context.Context, cartID int64, product string, price float64) (*entity.CartItem, error)
+	RemoveItem(ctx context.Context, cartID int64, cartItemID int64) error
+	GetCart(ctx context.Context, cartID int64) (*entity.Cart, error)
+	CalculatePrice(ctx context.Context, cartID int64) (*entity.PriceResponse, error)
 }
